Rename plugin 6 -> 7. Add test main
