+++ lib/migration/plugins/plugin7/vicadminProxyVarRename.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package plugin6
+package plugin7
 
 import (
 	"context"
